fix: adding bottom and topBar components
