+++ app/src/main/java/io/github/sustainow/MainActivity.kt
@@ -26,47 +26,15 @@ import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.activity.viewModels
 import androidx.annotation.RequiresApi
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.requiredSize
 import androidx.compose.foundation.layout.safeDrawingPadding
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.ArrowBack
-import androidx.compose.material.icons.automirrored.filled.ExitToApp
-import androidx.compose.material.icons.filled.AccountCircle
-import androidx.compose.material.icons.filled.Settings
-import androidx.compose.material3.DropdownMenu
-import androidx.compose.material3.DropdownMenuItem
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.NavigationBar
-import androidx.compose.material3.NavigationBarItem
-import androidx.compose.material3.NavigationBarItemColors
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
-import androidx.compose.material3.TopAppBar
-import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.unit.dp
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
 import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.navigation.NavDestination.Companion.hasRoute
-import androidx.navigation.NavDestination.Companion.hierarchy
 import androidx.navigation.NavGraph.Companion.findStartDestination
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
@@ -74,10 +42,8 @@ import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.navigation
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.toRoute
-import coil.compose.rememberAsyncImagePainter
 import dagger.hilt.android.AndroidEntryPoint
 import dagger.hilt.android.lifecycle.withCreationCallback
-import io.github.sustainow.domain.model.UserState
 import io.github.sustainow.presentation.theme.AppTheme
 import io.github.sustainow.presentation.ui.ConfigurationScreen
 import io.github.sustainow.presentation.ui.actions.CollectiveActionScreen
@@ -95,6 +61,8 @@ import io.github.sustainow.presentation.ui.actions.SearchCollectiveActionsScreen
 import io.github.sustainow.presentation.ui.SignUpScreen
 import io.github.sustainow.presentation.ui.actions.FormCollectiveActionScreen
 import io.github.sustainow.presentation.ui.actions.SubmitAction
+import io.github.sustainow.presentation.ui.components.BottomBar
+import io.github.sustainow.presentation.ui.components.TopBar
 import io.github.sustainow.presentation.viewmodel.CollectiveActionViewModel
 import io.github.sustainow.presentation.viewmodel.FormularyViewModel
 import io.github.sustainow.presentation.viewmodel.HomeViewModel
@@ -106,9 +74,7 @@ import io.github.sustainow.routes.HistoricCarbonFootprint
 import io.github.sustainow.routes.HistoricConsumeEnergy
 import io.github.sustainow.routes.HistoricConsumeWater
 import io.github.sustainow.routes.HistoricMainPage
-import io.github.sustainow.routes.routes
 import io.github.sustainow.service.auth.AuthService
-import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @AndroidEntryPoint
@@ -117,7 +83,6 @@ class MainActivity : ComponentActivity() {
     lateinit var authService: AuthService
 
     @RequiresApi(Build.VERSION_CODES.O)
-    @OptIn(ExperimentalMaterial3Api::class)
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         installSplashScreen()
@@ -127,14 +92,6 @@ class MainActivity : ComponentActivity() {
                 val navController = rememberNavController()
 
                 val userState by authService.user.collectAsState()
-
-                val context = LocalContext.current
-
-                var showUserMenu by remember {
-                    mutableStateOf(false)
-                }
-
-                val coroutineScope = rememberCoroutineScope()
 
                 val backStackEntry by navController.currentBackStackEntryAsState()
                 val currentScreen =
@@ -165,189 +122,13 @@ class MainActivity : ComponentActivity() {
                 Scaffold(
                     topBar = {
                         if (currentScreen != Login && currentScreen != SignUp) {
-                            TopAppBar(
-                                title = {
-                                    val logoResource =
-                                        painterResource(id = R.drawable.sustainow_logo_transparent)
-                                    Row(
-                                        modifier = Modifier.fillMaxWidth(),
-                                        verticalAlignment = Alignment.CenterVertically,
-                                    ) {
-                                        Spacer(modifier = Modifier.weight(1f)) // Adiciona espaço entre o botão de voltar e a logo
-                                        // Tornar a logo clicável
-                                        Image(
-                                            logoResource,
-                                            contentDescription = null,
-                                            modifier = Modifier.requiredSize(150.dp, 150.dp),
-                                        )
-                                        Spacer(modifier = Modifier.weight(1f)) // Centraliza a logo
-                                    }
-                                },
-                                colors =
-                                TopAppBarDefaults.topAppBarColors(
-                                    containerColor = MaterialTheme.colorScheme.tertiaryContainer,
-                                ),
-                                navigationIcon = {
-                                    if (canNavigateBack) {
-                                        IconButton(onClick = {
-                                            navController.popBackStack()
-                                        }) {
-                                            Icon(
-                                                imageVector = Icons.AutoMirrored.Filled.ArrowBack,
-                                                contentDescription = context.getString(R.string.back),
-                                            )
-                                        }
-                                    }
-                                },
-                                actions = {
-                                    when {
-                                        userState is UserState.Logged ->
-                                            if ((userState as UserState.Logged).user.profilePicture?.isNotEmpty() == true &&
-                                                (userState as UserState.Logged).user.profilePicture !== null
-                                            ) {
-                                                val painter =
-                                                    rememberAsyncImagePainter(
-                                                        model = (userState as UserState.Logged).user.profilePicture,
-                                                    )
-                                                IconButton(onClick = {
-                                                    showUserMenu = !showUserMenu
-                                                }) {
-                                                    Icon(
-                                                        painter = painter,
-                                                        contentDescription = context.getString(R.string.user_menu),
-                                                    )
-                                                    DropdownMenu(
-                                                        expanded = showUserMenu,
-                                                        onDismissRequest = {
-                                                            showUserMenu = false
-                                                        }) {
-                                                        DropdownMenuItem(
-                                                            text = { Text(context.getString(R.string.configuration)) },
-                                                            trailingIcon = {
-                                                                Icon(
-                                                                    Icons.Default.Settings,
-                                                                    contentDescription = context.getString(
-                                                                        R.string.configuration
-                                                                    ),
-                                                                )
-                                                            },
-                                                            onClick = {
-                                                                showUserMenu = false // Close the menu
-                                                                navController.navigate(Configuration) // Navigate to ConfigurationScreen
-                                                            },
-                                                        )
-                                                        DropdownMenuItem(
-                                                            text = { Text(context.getString(R.string.logout)) },
-                                                            trailingIcon = {
-                                                                Icon(
-                                                                    Icons.AutoMirrored.Filled.ExitToApp,
-                                                                    contentDescription = context.getString(
-                                                                        R.string.logout
-                                                                    ),
-                                                                )
-                                                            },
-                                                            onClick = {
-                                                                coroutineScope.launch {
-                                                                    authService.signOut()
-                                                                }
-                                                            },
-                                                        )
-                                                    }
-                                                }
-                                            } else {
-                                                IconButton(onClick = {
-                                                    showUserMenu = !showUserMenu
-                                                }) {
-                                                    Icon(
-                                                        Icons.Default.AccountCircle,
-                                                        contentDescription = context.getString(R.string.user_menu),
-                                                    )
-                                                    DropdownMenu(
-                                                        expanded = showUserMenu,
-                                                        onDismissRequest = {
-                                                            showUserMenu = false
-                                                        }) {
-                                                        DropdownMenuItem(
-                                                            text = { Text(context.getString(R.string.configuration)) },
-                                                            trailingIcon = {
-                                                                Icon(
-                                                                    Icons.Default.Settings,
-                                                                    contentDescription = context.getString(
-                                                                        R.string.configuration
-                                                                    ),
-                                                                )
-                                                            },
-                                                            onClick = {
-                                                                showUserMenu = false // Close the menu
-                                                                navController.navigate(Configuration) // Navigate to ConfigurationScreen
-                                                            },
-                                                        )
-                                                        DropdownMenuItem(
-                                                            text = { Text(context.getString(R.string.logout)) },
-                                                            trailingIcon = {
-                                                                Icon(
-                                                                    Icons.AutoMirrored.Filled.ExitToApp,
-                                                                    contentDescription = context.getString(
-                                                                        R.string.logout
-                                                                    ),
-                                                                )
-                                                            },
-                                                            onClick = {
-                                                                coroutineScope.launch {
-                                                                    authService.signOut()
-                                                                }
-                                                            },
-                                                        )
-                                                    }
-                                                }
-                                            }
-
-                                        else -> {
-                                        }
-                                    }
-                                },
-                            )
+                            TopBar(navController, authService, userState, canNavigateBack)
                         }
                     },
                     modifier = Modifier.safeDrawingPadding(),
                     bottomBar = {
-                        val currentDestination = backStackEntry?.destination
                         if (currentScreen != Login && currentScreen != SignUp) {
-                            NavigationBar(
-                                containerColor = MaterialTheme.colorScheme.surfaceContainerHighest,
-                            ) {
-                                routes().forEachIndexed { num, route ->
-                                    NavigationBarItem(
-                                        icon = {
-                                            Icon(
-                                                route.icon,
-                                                contentDescription = route.name,
-                                            )
-                                        },
-                                        label = {
-                                            Text(route.name)
-                                        },
-                                        // if the graph base route is anywhere in the current hierarchy
-                                        selected =
-                                        currentDestination?.hierarchy?.any {
-                                            it.hasRoute(route.content::class)
-                                        } == true,
-                                        colors =
-                                        NavigationBarItemColors(
-                                            selectedIconColor = MaterialTheme.colorScheme.primary,
-                                            selectedTextColor = MaterialTheme.colorScheme.primary,
-                                            selectedIndicatorColor = MaterialTheme.colorScheme.surface,
-                                            unselectedIconColor = MaterialTheme.colorScheme.onSurface,
-                                            unselectedTextColor = MaterialTheme.colorScheme.onSurface,
-                                            disabledIconColor = MaterialTheme.colorScheme.onSurfaceVariant,
-                                            disabledTextColor = MaterialTheme.colorScheme.onSurfaceVariant,
-                                        ),
-                                        onClick = {
-                                            navController.navigate(route.content)
-                                        },
-                                    )
-                                }
-                            }
+                            BottomBar(navController)
                         }
                     },
                 ) { innerPadding ->
